Introducing processor for TimeBasedValue
